version 2.6.6

+++ src/main/java/hal/studios/hpm/procedures/ShipwreckedraftRightClickedOnEntityProcedure.java
@@ -11,7 +11,7 @@ public class ShipwreckedraftRightClickedOnEntityProcedure {
 	public static void execute(LevelAccessor world, double x, double y, double z, Entity entity) {
 		if (entity == null)
 			return;
-		for (int index0 = 0; index0 < (int) (7); index0++) {
+		for (int index0 = 0; index0 < 7; index0++) {
 			if (world instanceof Level _level && !_level.isClientSide()) {
 				ItemEntity entityToSpawn = new ItemEntity(_level, x, y, z, new ItemStack(Blocks.OAK_LOG));
 				entityToSpawn.setPickUpDelay(10);
