+++ src/main/java/hal/studios/hpm/procedures/ShipwreckswashbucklerRightClickedOnEntityProcedure.java
@@ -11,7 +11,7 @@ public class ShipwreckswashbucklerRightClickedOnEntityProcedure {
 	public static void execute(LevelAccessor world, double x, double y, double z, Entity entity) {
 		if (entity == null)
 			return;
-		for (int index0 = 0; index0 < (int) (4); index0++) {
+		for (int index0 = 0; index0 < 4; index0++) {
 			if (world instanceof Level _level && !_level.isClientSide()) {
 				ItemEntity entityToSpawn = new ItemEntity(_level, x, y, z, new ItemStack(Blocks.OAK_PLANKS));
 				entityToSpawn.setPickUpDelay(10);
@@ -23,7 +23,7 @@ public class ShipwreckswashbucklerRightClickedOnEntityProcedure {
 				_level.addFreshEntity(entityToSpawn);
 			}
 		}
-		for (int index1 = 0; index1 < (int) (2); index1++) {
+		for (int index1 = 0; index1 < 2; index1++) {
 			if (world instanceof Level _level && !_level.isClientSide()) {
 				ItemEntity entityToSpawn = new ItemEntity(_level, x, y, z, new ItemStack(Blocks.OAK_LOG));
 				entityToSpawn.setPickUpDelay(10);
